# openapi.yaml
openapi: 3.0.3
info:
  title: Project Volusia — Contribution & Data API
  description: |
    The machine-readable interface for querying Volusia County open data
    and submitting contributions to the Project Volusia knowledge system.

    Three main surfaces:
    1. **Data API** — read-only access to published indicators, datasets, charts, and analysis
    2. **Contribution API** — submit new data, analysis, tools, maps, and community knowledge
    3. **Health API** — system health, status, and data freshness monitoring

    Full documentation: see AGENTIC_CONTRIBUTION_STRATEGY.md
  version: "2026-09-06"
  contact:
    name: Project Volusia Community Liaison
    email: community@project-volusia.org
  license:
    name: MIT
servers:
  - url: https://api.project-volusia.org/api/v1
    description: Production
  - url: http://localhost:8789/api/v1
    description: Local development
...
paths:
...
  /contributions:
    post:
      tags: [Contributions]
      summary: Submit a new contribution
      description: |
        Submit a new contribution to the knowledge system. Anonymous submissions
        are allowed unless VOLUSIA_API_KEYS is set. Rate limited to 10 requests
        per 60 seconds per client (configurable via VOLUSIA_RATE_LIMIT and
        VOLUSIA_RATE_WINDOW env vars).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContributionSubmission" }
      responses:
        "201":
          description: Contribution accepted for review
          headers:
            X-RateLimit-Limit:
              schema: { type: integer }
              description: Max requests per window
            X-RateLimit-Remaining:
              schema: { type: integer }
              description: Remaining requests in current window
          content:
            application/json:
              schema:
                type: object
                properties:
                  submission_id: { type: string }
                  status:
                    type: string
                    enum: [queued, under_review, approved, rejected, needs_revision]
                  submitted_at: { type: string, format: date-time }
                  acknowledged_at: { type: string, format: date-time }
                  estimated_review_by: { type: string, format: date-time }
                  reviewer: { type: string }
                  anonymous: { type: boolean }
                  message: { type: string }
        "400":
          description: Validation error
        "401":
          description: Missing or invalid API key
        "429":
          description: Rate limit exceeded
          headers:
            Retry-After:
              schema: { type: number }
              description: Seconds until next request allowed
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: object
                    properties:
                      error: { type: string }
                      retry_after_seconds: { type: number }
                      limit: { type: integer }
                      window_seconds: { type: integer }

    get:
      tags: [Contributions]
      summary: List recent contributions
      description: List recent submissions (for CGB triage)
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, minimum: 1, maximum: 100 }
        - name: offset
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
        - name: status
          in: query
          schema: { type: string }
          description: Filter by status (queued, under_review, approved, rejected, needs_revision)
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  submissions:
                    type: array
                    items:
                      type: object
                      properties:
                        submission_id: { type: string }
                        contribution_type: { type: string }
                        status: { type: string }
                        submitted_at: { type: string }
                        reviewer: { type: string }
                        author_name: { type: string }
  /contributions/{submission_id}:
    get:
      tags: [Contributions]
      summary: Get contribution status
      parameters:
        - name: submission_id
          in: path
          required: true
          schema:
            type: string
    patch:
      tags: [Contributions]
      summary: Update contribution status
      description: Update a submission's status (CGB triage use)
      parameters:
        - name: submission_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [queued, under_review, approved, rejected, needs_revision]
                decision: { type: string }
                reviewer: { type: string }
      responses:
        "200":
          description: Submission updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  submission_id: { type: string }
                  status: { type: string }
                  decision: { type: string }
                  reviewer: { type: string }
                  acknowledged_at: { type: string }
                  message: { type: string }
        "400":
          description: Invalid status value
        "404":
          description: Submission not found
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    ContributionSubmission:
      type: object
      required: [contribution_type, content]
      properties:
        contribution_type:
          type: string
          enum: [data_source, analysis, tool, map, report, community, social_media, educational, direct]
        content:
          description: The contribution payload (structure varies by type)
          oneOf:
            - type: string
            - type: object
        author_name: { type: string }
        author_email: { type: string, format: email }
        idempotency_key:
          type: string
          description: Optional key for safe retry (returns existing submission on duplicate)
...